+++ .isort.cfg
@@ -1,3 +1,5 @@
 [settings]
-known_first_party = libmozevent
+known_first_party = conftest,libmozevent
 known_third_party = aioamqp,aiohttp,aioredis,hglib,libmozdata,logbook,pytest,responses,setuptools,structlog,taskcluster
+force_single_line = True
+line_length = 159
